+++ src/control_plane/shared_infrastructure/mysql/mod.rs
@@ -6,6 +6,7 @@ mod mysql_shared_instance_plan;
 mod mysql_shared_instance_plan_options;
 mod plan_mysql_project_resources;
 mod provision_mysql_logical_resource;
+mod reconcile_mysql_project_resources;
 
 pub(crate) use mysql_flavor::MySqlFlavor;
 pub(crate) use mysql_logical_resource_plan::MySqlLogicalResourcePlan;
@@ -15,3 +16,4 @@ pub(crate) use mysql_shared_instance_plan::MySqlSharedInstancePlan;
 pub(crate) use mysql_shared_instance_plan_options::MySqlSharedInstancePlanOptions;
 pub(crate) use plan_mysql_project_resources::plan_mysql_project_resources;
 pub(crate) use provision_mysql_logical_resource::provision_mysql_logical_resource;
+pub(crate) use reconcile_mysql_project_resources::reconcile_mysql_project_resources;